Crear VideoObjectMother y VideosObjectMother
