+++ tests/Mooc/Videos/Infrastructure/VideoRepositoryRandomFake.php
@@ -2,16 +2,13 @@
 
 namespace CodelyTv\Tests\Mooc\Videos\Infrastructure;
 
-use CodelyTv\Mooc\Shared\Domain\Courses\CourseId;
-use CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl;
 use CodelyTv\Mooc\Videos\Domain\Video;
 use CodelyTv\Mooc\Videos\Domain\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoRepository;
 use CodelyTv\Mooc\Videos\Domain\Videos;
-use CodelyTv\Mooc\Videos\Domain\VideoTitle;
-use CodelyTv\Mooc\Videos\Domain\VideoType;
 use CodelyTv\Shared\Domain\Criteria\Criteria;
-use CodelyTv\Shared\Domain\ValueObject\Uuid;
+use CodelyTv\Tests\Mooc\Videos\Domain\VideoObjectMother;
+use CodelyTv\Tests\Mooc\Videos\Domain\VideosObjectMother;
 
 class VideoRepositoryRandomFake implements VideoRepository
 {
@@ -22,27 +19,11 @@ class VideoRepositoryRandomFake implements VideoRepository
 
     public function search(VideoId $id): ?Video
     {
-        return self::getRandomVideoWithId($id->value());
+        return VideoObjectMother::getRandomVideoWithId($id->value());
     }
 
     public function searchByCriteria(Criteria $criteria): Videos
     {
-        return self::getRandomVideos();
-    }
-
-    public static function getRandomVideoWithId(string $idValue): Video
-    {
-        return new Video(
-            new VideoId((new Uuid($idValue))->value()),
-            VideoType::random(),
-            new VideoTitle('foobar'),
-            new VideoUrl('https://www.foobar.com/foo'),
-            new CourseId(CourseId::random()->value())
-        );
-    }
-
-    public static function getRandomVideos(): Videos
-    {
-        return new Videos([self::getRandomVideoWithId(Uuid::random()->value())]);
+        return VideosObjectMother::getRandomVideos();
     }
 }
